Shaders: fixed freestyle lanes for RB2 tales.
git-svn-id: http://fofix.googlecode.com/svn/MFH-Mod/trunk@1292 48f48816-a574-47fe-9a59-6914b3aef992

+++ src/Guitar.py
@@ -1420,8 +1420,9 @@ class Guitar:
             color = (c1 + c1*2.0*tailGlow, c2 + c2*2.0*tailGlow, c3 + c3*2.0*tailGlow, c4*0.6 + c4*0.4*tailGlow)    #MFH - this fades inactive tails' color darker                    
             
           tailcol = (color)
-          
-        if self.theme == 2 and big and tailOnly and Shader.list.enable("tail"):
+
+        if self.theme == 2 and freestyleTail == 0 and big and tailOnly and Shader.list.enable("tail"):
+        
           if kill and self.killfx == 0:
             Shader.list.setVar("color",(1.0,1.0,1.0,1.0))
             Shader.list.setVar("height",0.05)
